+++ rtic-macros/ui/shared-pub.rs
@@ -1,9 +1,0 @@
-#![no_main]
-
-#[rtic_macros::mock_app(device = mock)]
-mod app {
-    #[shared]
-    struct Shared {
-        pub x: u32,
-    }
-}