+++ rtic-time/CHANGELOG.md
@@ -9,8 +9,12 @@ For each category, *Added*, *Changed*, *Fixed* add new entries at the top!
 
 ### Added
 
+- `should_dequeue` to the `Monotonic` trait to handle bugged timers
+
 ### Changed
 
 ### Fixed
+
+- If the queue was non-empty and a new instant was added that was earlier than `head`, then the queue would no pend the monotonic handler. This would cause the new `head` to be dequeued at the wrong time.
 
 ## [v1.0.0] - 2023-xx-xx