+++ rtic/CHANGELOG.md
@@ -13,6 +13,10 @@ For each category, *Added*, *Changed*, *Fixed* add new entries at the top!
 
 ### Changed
 
+- `cortex-m` set as an optional dependency
+- Moved `cortex-m`-related utilities from `rtic/lib.rs` to `rtic/export.rs`
+- Make async task priorities start at 0, instead of 1, to always start at the lowest priority
+
 ## [v1.1.4] - 2023-02-26
 
 ### Added